Add service and device selectors to notifications.
Previously notifications were only filtered by characteristics. This however results in problems when working with devices that
have multiple of the same characteristics in different services.

For example: If a luminaire consisted of two lamps and each lamp has the characteristic 'brightness' then notifications for one would always show both.